Add Shadow Sceptre tracking
Added tracking for shadow barrage with the shadow sceptre.

Added params to ensure shadow barrage is a one-time reduction.

Added ashadow sceptre infobox, similar to vulernability (except the icon is the sceptre, because the barrage was ugly).

Refactored some of the onGraphicChanged methods to be a bit cleaner given there are two possible outcomes here.